Refactor import paths to use `@/shadcn/*` for UI components
