+++ web-app/eslint.config.mjs
@@ -7,12 +7,15 @@ const __dirname = dirname(__filename);
 
 const compat = new FlatCompat({
   baseDirectory: __dirname,
+  recommendedConfig: {
+    extends: ["eslint:recommended", "plugin:react/recommended"],
+  },
 });
 
 const eslintConfig = [
   ...compat.config({
     extends: ["next/core-web-vitals", "next/typescript", "prettier"],
-    plugins: ["simple-import-sort", "prettier"],
+    plugins: ["simple-import-sort", "import", "prettier", "unused-imports"],
     ignorePatterns: [
       "**/dist/**",
       "**/build/**",
@@ -24,6 +27,14 @@ const eslintConfig = [
       "simple-import-sort/imports": "error",
       "simple-import-sort/exports": "error",
       "prettier/prettier": "error",
+      "import/first": "error",
+      "import/newline-after-import": "error",
+      "import/no-duplicates": "error",
+      "unused-imports/no-unused-imports": "error",
+    },
+    parserOptions: {
+      sourceType: "module",
+      ecmaVersion: "latest",
     },
   }),
 ];
