+++ web-app/package.json
@@ -78,6 +78,7 @@
     "eslint-config-prettier": "^10.1.1",
     "eslint-plugin-prettier": "^5.2.3",
     "eslint-plugin-simple-import-sort": "^12.1.1",
+    "eslint-plugin-unused-imports": "^4.1.4",
     "jest": "29.7.0",
     "jest-environment-jsdom": "29.7.0",
     "postcss": "^8.5.3",
