+++ web-app/src/app/(app)/components/sidebar/components/agents-list.tsx
@@ -1,4 +1,4 @@
-import { ScrollArea, ScrollBar } from "@/components/ui/scroll-area";
+import { ScrollArea, ScrollBar } from "@/shadcn/scroll-area";
 import {
   SidebarGroup,
   SidebarGroupContent,
@@ -6,7 +6,7 @@ import {
   SidebarMenu,
   SidebarMenuButton,
   SidebarMenuItem,
-} from "@/components/ui/sidebar";
+} from "@/shadcn/sidebar";
 
 const agentsGroups = [
   {
