+++ web-app/src/app/(landing)/page/how-it-works.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import Link from "next/link";
 import { useTranslations } from "next-intl";
 
-import { Button } from "@/components/ui/button";
+import { Button } from "@/shadcn/button";
 
 export default function HowItWorks() {
   const t = useTranslations("Landing.Page.HowItWorks");
